prepare the entrypoint
